+++ src/packages/popup/index.js
@@ -1,0 +1,7 @@
+import Popup from "./popup.vue";
+
+Popup.install = function(Vue) {
+  Vue.component(Popup.name, Popup);
+};
+
+export default Popup;