+++ types/nutui.d.ts
@@ -63,3 +63,4 @@ export declare class TextInput extends UIComponent {}
 export declare class Avatar extends UIComponent {}
 export declare class Infiniteloading extends UIComponent {}
 export declare class Lazyload extends UIComponent {}
+export declare class Elevator extends UIComponent {}